Refactor project:info-update into trait.
